nginx-0.3.53-RELEASE import
    *) Change: the "add_header" directive adds the string to 204, 301, and
       302 responses.

    *) Feature: the "server" directive in the "upstream" context supports
       the "weight" parameter.

    *) Feature: the "server_name" directive supports the "*" wildcard.

    *) Feature: nginx supports the request body size more than 2G.

    *) Bugfix: if a client was successfully authorized using "satisfy_any
       on", then anyway the message "access forbidden by rule" was written
       in the log.

    *) Bugfix: the "PUT" method may erroneously not create a file and
       return the 409 code.

    *) Bugfix: if the IMAP/POP3 backend returned an error, then nginx
       continued proxying anyway.

+++ src/http/ngx_http_request.h
@@ -192,7 +192,7 @@ typedef struct {
     ngx_array_t                       cookies;
 
     size_t                            host_name_len;
-    ssize_t                           content_length_n;
+    off_t                             content_length_n;
     time_t                            keep_alive_n;
 
     unsigned                          connection_type:2;
@@ -250,7 +250,7 @@ typedef struct {
     ngx_temp_file_t                  *temp_file;
     ngx_chain_t                      *bufs;
     ngx_buf_t                        *buf;
-    size_t                            rest;
+    off_t                             rest;
     ngx_chain_t                      *to_write;
     ngx_http_client_body_handler_pt   post_handler;
 } ngx_http_request_body_t;
@@ -366,7 +366,7 @@ struct ngx_http_request_s {
     /* used to learn the Apache compatible response length without a header */
     size_t                            header_size;
 
-    size_t                            request_length;
+    off_t                             request_length;
 
     void                            **err_ctx;
     ngx_uint_t                        err_status;
